itj-prod2/itj/requirements.txt
asn1crypto==0.24.0
cement==2.8.2
certifi==2019.3.9
cffi==1.12.2
chardet==3.0.4
colorama==0.3.9
Django==2.1.7
django-appconf==1.0.3
django-ckeditor==5.6.1
django-favicon==0.1.3
django-js-asset==1.2.2
docutils==0.14
future==0.16.0
idna==2.7
jmespath==0.9.4
pathspec==0.5.9
Pillow==5.4.1
protobuf==3.7.1
psycopg2==2.7.7
pycparser==2.19
python-dateutil==2.8.0
pytz==2018.9
PyYAML==3.13
requests==2.20.1
semantic-version==2.5.0
six==1.11.0
termcolor==1.1.0
urllib3==1.24.2
wcwidth==0.1.7
django-el-pagination==3.2.4
django-storages==1.7.1
boto3==1.9.146
botocore==1.12.146
s3transfer==0.2.0
beautifulsoup4==4.7.1
selenium==3.141.0
